## Proposed changes

The JSON returned by the past API was not formatted.
before 
```
curl -X GET -u 'root:' http://127.0.0.1:8030/api/profile?query_id=48bdf6d75dbb46c9-998b9c0368f4561f >> a1.profile

{"msg":"success","code":0,"data":{"profile":"Summary:\n   - Profile ID: 48bdf6d75dbb46c9-998b9c0368f4561f\n   - Task Type: QUERY\n   - Start Time: 2023-12-20 11:09:41\n   - End Time: 2023-12-20 11:09:45\n   - Total: 3s680ms\n   - Task State: EOF\n   - User: root\n   - Default Db: tpcds\n   - Sql Statement: with customer_total_return as\n(select sr_customer_sk as ctr_customer_sk\n,sr_store_sk as ctr_store_sk\n,sum(SR_FEE) as ctr_total_return\nfrom store_returns\n,date_dim\nwhere sr_returned_date_sk = d_date_sk\nand d_year =2000\ngroup by sr_customer_sk\n,sr_store_sk)\n select  c_customer_id\nfrom customer_total_return ctr1\n,store\n,customer\nwhere ctr1.ctr_total_return > (select avg(ctr_total_return)*1.2\nfrom customer_total_return ctr2\nwhere ctr1.ctr_store_sk = ctr2.ctr_store_sk)\nand s_store_sk = ctr1.ctr_store_sk\nand s_state = 'TN'\nand ctr1.ctr_customer_sk = c_customer_sk\norder by c_customer_id\nlimit 100\nExecution Summary:\n   - Nereids Analysis Time: 321ms\n   - Nereids Rewrite Time: 433ms\n   - Nereids Optimize Time: 218ms\n   - Nereids Translate Time: 39ms\n   - Workload Group: \n   - Ana
```

now

```
curl -X GET -u 'root:' http://127.0.0.1:8030/api/profile/text?query_id=48bdf6d75dbb46c9-998b9c0368f4561f >> a2.profile

Summary:
   - Profile ID: 48bdf6d75dbb46c9-998b9c0368f4561f
   - Task Type: QUERY
   - Start Time: 2023-12-20 11:09:41
   - End Time: 2023-12-20 11:09:45
   - Total: 3s680ms
   - Task State: EOF
   - User: root
   - Default Db: tpcds
   - Sql Statement: with customer_total_return as
(select sr_customer_sk as ctr_customer_sk

```

<!--Describe your changes.-->

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...